refactored event data management
